+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.5.2] - 16/03/2021    
+
+- Implement singleton cache for events being processed
+- Add JUnit tests
+
 ## [1.5.1] - 03/02/2021      
 
 - Add configuration number of threads and limit rate in files processing